Avant modif heredoc

+++ src/pipex/pipex.c
@@ -6,7 +6,7 @@
 /*   By: lgirault <lgirault@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/13 16:37:23 by lgirault          #+#    #+#             */
-/*   Updated: 2023/05/09 11:58:04 by lgirault         ###   ########.fr       */
+/*   Updated: 2023/05/10 15:38:15 by lgirault         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,6 +62,7 @@ int	parent(t_cmd *cmd, t_ms *ms)
 			cmd->exit_status = WEXITSTATUS(status);
 		cmd->i--;
 	}
+	signal(SIGINT, handler_sigint);
 	free(cmd->pipefd);
 	free(cmd->pid);
 	if (ms->builtin_code == 0)
@@ -76,24 +77,53 @@ int	pipex(t_cmd_lst *cmd_lst, t_ms *ms)
 
 	init_tab(&cmd, cmd_lst);
 	init(&cmd, cmd_lst, (*ms).env);
+	// init_tab_closefile(&cmd, cmd_lst, ms);
+	// open_infile_heredoc(cmd, cmd_lst, ms);
+	///check heredoc ici si pas ctrl c lance boucle sinon on quitte en supprimant le heredoc
 	while (cmd_lst != NULL)
 	{
+		// printf("%s\n", cmd_lst->cmd_option[0]);
+		// if (cmd_lst->infile_name != NULL)
+		// {
+		// 	printf("%s\n", cmd_lst->infile_name[0]);
+		// }
+		// if (cmd_lst->limit_mode != NULL)
+		// {
+		// 	printf("%d\n", cmd_lst->limit_mode[0]);
+		// }
+		//printf("%d\n", cmd_lst->limit_mode[0]);
 		init_tab_closefile(&cmd, cmd_lst, ms);
 		for_open(cmd_lst, &cmd, ms);
-		if (ms->lock_cat != 1)
-			ms->cat_grep = check_cat_grep(cmd_lst, ms);
-		cmd.pid[cmd.i] = fork();
-		if (cmd.pid[cmd.i] == -1)
-			error_management(2, &cmd);
-		if (cmd.pid[cmd.i] == 0)
-			child(&cmd, (*ms).env, cmd_lst, ms);
-		close_fichier(cmd, cmd_lst, ms->env);
-		only_last(cmd_lst, ms, &cmd, 1);
-		if (access("/tmp/.file_temp.txt", F_OK) == 0)
-			if (cmd_lst->limit_mode[tab_len(cmd_lst->infile_name)] == 2)
-				if (unlink("/tmp/.file_temp.txt") == -1)
-					error_management(7, &cmd);
-		clear_lst(&cmd_lst);
+		if (ms->builtin_code != 130)
+		{
+			cmd.pid[cmd.i] = fork();
+			if (cmd.pid[cmd.i] == -1)
+				error_management(2, &cmd);
+			else
+				signal(SIGINT, SIG_IGN);
+			if (cmd.pid[cmd.i] == 0)
+			{
+				sig_for_child();
+				child(&cmd, (*ms).env, cmd_lst, ms);
+			}
+			close_fichier(cmd, cmd_lst, ms->env);
+			only_last(cmd_lst, ms, &cmd, 1);
+			if (access("/tmp/.file_temp.txt", F_OK) == 0)
+				if (cmd_lst->limit_mode[tab_len(cmd_lst->infile_name)] == 2)
+					if (unlink("/tmp/.file_temp.txt") == -1)
+						error_management(7, &cmd);
+			clear_lst(&cmd_lst);
+		}
+		else
+		{
+			close_fichier(cmd, cmd_lst, ms->env);
+			if (access("/tmp/.file_temp.txt", F_OK) == 0)
+				if (cmd_lst->limit_mode[tab_len(cmd_lst->infile_name)] == 2)
+					if (unlink("/tmp/.file_temp.txt") == -1)
+						error_management(7, &cmd);
+			lstclear(&cmd_lst);
+			break ;
+		}
 		cmd.i++;
 	}
 	return (parent(&cmd, ms));
